+++ src/ProjectTemplates/Web.ProjectTemplates/content/RazorComponentsWeb-CSharp/Data/WeatherForecast.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace RazorComponentsWeb_CSharp.Services
+namespace RazorComponentsWeb_CSharp.Data
 {
     public class WeatherForecast
     {
