+++ src/ProjectTemplates/Web.ProjectTemplates/content/RazorComponentsWeb-CSharp/Pages/FetchData.razor
@@ -1,5 +1,5 @@
 @page "/fetchdata"
-@using RazorComponentsWeb_CSharp.Services
+@using RazorComponentsWeb_CSharp.Data
 @inject WeatherForecastService ForecastService
 
 <h1>Weather forecast</h1>
